move token & price creation into method
